// src/enums/calendarBooking.ts
export enum CalendarBookingActionTypes {
    FETCH_SELECTED_DAY_BOOKINGS = '@@calendarBooking/FETCH_SELECTED_DAY_BOOKINGS',
    FETCH_SELECTED_DAY_BOOKINGS_SUCCESS = '@calendarBooking/FETCH_SELECTED_DAY_BOOKINGS_SUCCESS',
    FETCH_SELECTED_DAY_BOOKINGS_FAILURE = '@calendarBooking/FETCH_SELECTED_DAY_BOOKINGS_FAILURE',
    FETCH_TUTOR_SUBJECTS_REQUEST = '@@calendarBooking/FETCH_TUTOR_SUBJECTS_REQUEST',
    FETCH_TUTOR_SUBJECTS_SUCCESS = '@@calendarBooking/FETCH_TUTOR_SUBJECTS_SUCCESS',
    FETCH_TUTOR_SUBJECTS_FAILURE = '@@calendarBooking/FETCH_TUTOR_SUBJECTS_FAILURE',
    SELECT_DATE = '@@calendarBooking/SELECT_DATE',
    CLEAR_DATA = '@@calendarBooking/CLEAR_DATA',
    CREATE_BOOKING = '@@calendarBooking/CREATE_BOOKING',
}